move bookmarks utils into bookmarks package add getConfig to pluginapi
