+++ server/utils/utils.go
@@ -1,17 +1,10 @@
 package utils
 
 import (
-	"fmt"
 	"regexp"
 
 	"github.com/mattermost/mattermost-server/v5/plugin"
 )
-
-// getPermaLink returns a link to a postID
-func getPermaLink(postID string) string {
-	url := "https://myhost.com"
-	return fmt.Sprintf("%v/_redirect/pl/%v", url, postID)
-}
 
 // GetPostIDFromLink extracts a PostID from a link
 func GetPostIDFromLink(s string) string {
@@ -20,12 +13,6 @@ func GetPostIDFromLink(s string) string {
 		return r.FindStringSubmatch(s)[1]
 	}
 	return s
-}
-
-// GetIconLink returns a markdown link to a postID including a :link: icon
-func GetIconLink(postID string) string {
-	iconLink := fmt.Sprintf("[:link:](%s)", getPermaLink(postID))
-	return iconLink
 }
 
 type API struct {
